refactor(@angular/cli): extract update migration logic into separate module
Extracts the migration execution and related helper logic from `cli.ts`
into a new module `utilities/migration.ts` within the update command.
